+++ SuiviRegime/src/s6/suiviRegime/modele/SportConseil.java
@@ -3,7 +3,7 @@ package s6.suiviRegime.modele;
 public class SportConseil extends BaseModele{
 	
 	private Sport sport;
-	private float rythme;
+	private int rythme;
 	private String details;
 	
 	public SportConseil() {}
@@ -12,13 +12,13 @@ public class SportConseil extends BaseModele{
 		super(id);
 	}
 
-	public SportConseil(int id, int idSport, String libelle, String activite, float rythme, String details) {
+	public SportConseil(int id, int idSport, String libelle, String activite, int rythme, String details) {
 		this.setId(id);
 		this.setSport(new Sport(idSport, libelle, activite));
 		this.setRythme(rythme);
 		this.setDetails(details);
 	}
-	public SportConseil(int id, Sport sport, float rythme, String details) {
+	public SportConseil(int id, Sport sport, int rythme, String details) {
 		this.setId(id);
 		this.setSport(sport);
 		this.setRythme(rythme);
@@ -32,10 +32,10 @@ public class SportConseil extends BaseModele{
 		this.sport = sport;
 	}
 
-	public float getRythme() {
+	public int getRythme() {
 		return rythme;
 	}
-	public void setRythme(float rythme) {
+	public void setRythme(int rythme) {
 		this.rythme = rythme;
 	}
 	public void setRythme(String rythme)throws Exception{
